+++ integrations/checkmarx-one/CHANGELOG.md
@@ -7,6 +7,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+## 0.3.4-dev (2025-09-17)
+
+
+### Improvements
+
+- Add ResourceConfig to list of resource in integration config file
+
+
+## 0.3.3-dev (2025-09-17)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.28.7
+
+
+## 0.3.2-dev (2025-09-16)
+
+
+### Features
+
+- Added webhook processor for project creation events
+
+
+## 0.3.1-dev (2025-09-16)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.28.5
+
+
 ## 0.3.0-dev (2025-09-10)